refactor:  hot게시물까지 진행
