+++ src/app/api/internal/lounge.ts
@@ -5,7 +5,7 @@ import type { GetPostsParams, GetPostsResponse, Post } from "@/types";
  * isThread 필터링 후에도 정확히 size개를 채울 때까지
  * 백엔드를 연속 호출하는 BFF 전용 fetcher
  */
-export async function fetchLoungePostsPage(
+export async function getLoungePostsPageBFF(
   params: GetPostsParams & { size: number },
 ): Promise<GetPostsResponse> {
   const { size, cursor: initialCursor, ...rest } = params;
